fix(expression): count the items of a dynamic IN in a width no list exceeds
The item count was narrowed to u16, so a list of 65,536 items compiled to InList(0), which took the last item for the value, answered false and left the items on the stack. The count is now u32, and a list beyond it is refused at compile time instead of being narrowed.

+++ src/executor/expression/ops.rs
@@ -617,7 +617,7 @@ pub enum Op {
 
     /// IN over items evaluated at run time; the value is judged once
     /// Stack: [value, item_1, ..., item_n] -> [bool]
-    InList(u16), // item count
+    InList(u32), // item count
 
     /// BETWEEN check: value BETWEEN low AND high
     /// Stack: [value, low, high] -> [bool]
